removed warnings
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@12533 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: fe4680c786d5e960684730c040ae94a01717ca79

+++ mapping/point_cloud_mapping/src/cloud_downsampler.cpp
@@ -101,7 +101,7 @@ class CloudDownsampler
     // Callback
     void cloud_cb ()
     {
-      ROS_INFO ("Received %d data points.", cloud_.pts.size ());
+      ROS_INFO ("Received %d data points.", (int)cloud_.pts.size ());
       if (cloud_.pts.size () == 0)
         return;
 
@@ -113,7 +113,7 @@ class CloudDownsampler
 
       gettimeofday (&t2, NULL);
       double time_spent = t2.tv_sec + (double)t2.tv_usec / 1000000.0 - (t1.tv_sec + (double)t1.tv_usec / 1000000.0);
-      ROS_INFO ("Cloud downsampled in %g seconds. Number of points left: %d.", time_spent, cloud_down_.pts.size ());
+      ROS_INFO ("Cloud downsampled in %g seconds. Number of points left: %d.", time_spent, (int)cloud_down_.pts.size ());
 
       node_.publish ("cloud_downsampled", cloud_down_);
     }
